@@ -9,7 +9,7 @@ Supports:
 """
 
 from pathlib import Path
-from typing import Any, Dict, Optional, Union
+from typing import Any
 
 import yaml
 from pydantic import ValidationError
@@ -34,7 +34,7 @@ from ced_ml.config.schema import (
 )
 
 
-def load_yaml(file_path: Union[str, Path]) -> Dict[str, Any]:
+def load_yaml(file_path: str | Path) -> dict[str, Any]:
     """Load configuration from YAML file."""
     file_path = Path(file_path)
     if not file_path.exists():
@@ -46,7 +46,68 @@ def load_yaml(file_path: Union[str, Path]) -> Dict[str, Any]:
     return config_dict or {}
 
 
-def apply_overrides(config_dict: Dict[str, Any], overrides: list[str]) -> Dict[str, Any]:
+def resolve_paths_relative_to_config(
+    config_dict: dict[str, Any], config_file: Path
+) -> dict[str, Any]:
+    """
+    Resolve relative paths in config dict relative to config file directory.
+
+    Args:
+        config_dict: Configuration dictionary (may contain Path-like strings)
+        config_file: Path to the config file
+
+    Returns:
+        Config dict with relative paths resolved
+
+    Note:
+        Only resolves paths that:
+        - Are relative (not absolute)
+        - Exist as keys containing "file", "dir", or "path" (case-insensitive)
+        - Point to files/dirs that can be resolved relative to config dir
+    """
+    config_dir = config_file.resolve().parent
+    resolved_dict = config_dict.copy()
+
+    def resolve_value(value: Any) -> Any:
+        """Recursively resolve Path-like values."""
+        if isinstance(value, dict):
+            return {k: resolve_value(v) for k, v in value.items()}
+        elif isinstance(value, list):
+            return [resolve_value(v) for v in value]
+        elif isinstance(value, str):
+            # Check if this looks like a path (not absolute)
+            path = Path(value)
+            if not path.is_absolute() and len(path.parts) > 0:
+                # Try resolving relative to config dir
+                resolved = config_dir / path
+                # Only replace if resolved path exists or value is clearly a path
+                if resolved.exists() or "/" in value or "\\" in value:
+                    return str(resolved)
+        return value
+
+    # Only resolve keys that look like paths
+    path_like_keys = [
+        "infile",
+        "outdir",
+        "split_dir",
+        "results_dir",
+        "model_artifact",
+        "holdout_idx",
+    ]
+
+    for key, val in resolved_dict.items():
+        if key in path_like_keys:
+            resolved_dict[key] = resolve_value(val)
+        elif isinstance(val, dict):
+            # Recursively resolve nested dicts
+            for nested_key, nested_val in val.items():
+                if nested_key in path_like_keys:
+                    val[nested_key] = resolve_value(nested_val)
+
+    return resolved_dict
+
+
+def apply_overrides(config_dict: dict[str, Any], overrides: list[str]) -> dict[str, Any]:
     """
     Apply CLI overrides to config dictionary.
 
@@ -70,6 +131,12 @@ def apply_overrides(config_dict: Dict[str, Any], overrides: list[str]) -> Dict[s
         "toprisk_fracs",
     }
 
+    # Keys that should always be strings (not parsed as int/float)
+    STRING_KEYS = {
+        "run_id",
+        "run_name",
+    }
+
     for override in overrides:
         if "=" not in override:
             raise ValueError(f"Invalid override format: {override}. Expected 'key=value'")
@@ -87,20 +154,26 @@ def apply_overrides(config_dict: Dict[str, Any], overrides: list[str]) -> Dict[s
         # Parse value (try int, float, bool, then string)
         final_key = keys[-1]
         force_list = final_key in LIST_KEYS
-        value = _parse_value(value_str, force_list=force_list)
+        force_string = final_key in STRING_KEYS
+        value = _parse_value(value_str, force_list=force_list, force_string=force_string)
         target[final_key] = value
 
     return config_dict
 
 
-def _parse_value(value_str: str, force_list: bool = False) -> Any:
+def _parse_value(value_str: str, force_list: bool = False, force_string: bool = False) -> Any:
     """
     Parse string value to appropriate Python type.
 
     Args:
         value_str: String to parse
         force_list: If True, always return a list (for comma-separated or single values)
+        force_string: If True, always return a string (skip int/float parsing)
     """
+    # Force string if requested
+    if force_string:
+        return value_str
+
     # Boolean
     if value_str.lower() in ("true", "yes", "1"):
         return [True] if force_list else True
@@ -149,8 +222,8 @@ def _parse_value(value_str: str, force_list: bool = False) -> Any:
 
 
 def load_splits_config(
-    config_file: Optional[Union[str, Path]] = None,
-    overrides: Optional[list[str]] = None,
+    config_file: str | Path | None = None,
+    overrides: list[str] | None = None,
 ) -> SplitsConfig:
     """
     Load splits configuration from file and CLI overrides.
@@ -167,7 +240,12 @@ def load_splits_config(
 
     # Load from file if provided
     if config_file is not None:
-        file_config = load_yaml(config_file)
+        config_file_path = Path(config_file)
+        file_config = load_yaml(config_file_path)
+
+        # Resolve relative paths relative to config file directory
+        file_config = resolve_paths_relative_to_config(file_config, config_file_path)
+
         config_dict.update(file_config)
 
     # Apply CLI overrides
@@ -182,8 +260,8 @@ def load_splits_config(
 
 
 def load_training_config(
-    config_file: Optional[Union[str, Path]] = None,
-    overrides: Optional[list[str]] = None,
+    config_file: str | Path | None = None,
+    overrides: list[str] | None = None,
 ) -> TrainingConfig:
     """
     Load training configuration from file and CLI overrides.
@@ -210,7 +288,11 @@ def load_training_config(
 
     # Load from file if provided
     if config_file is not None:
-        file_config = load_yaml(config_file)
+        config_file_path = Path(config_file)
+        file_config = load_yaml(config_file_path)
+
+        # Resolve relative paths relative to config file directory
+        file_config = resolve_paths_relative_to_config(file_config, config_file_path)
 
         # Deep merge nested dicts
         for key, value in file_config.items():
@@ -231,8 +313,8 @@ def load_training_config(
 
 
 def load_aggregate_config(
-    config_file: Optional[Union[str, Path]] = None,
-    overrides: Optional[list[str]] = None,
+    config_file: str | Path | None = None,
+    overrides: list[str] | None = None,
 ) -> AggregateConfig:
     """
     Load aggregate configuration from file and CLI overrides.
@@ -247,7 +329,11 @@ def load_aggregate_config(
     config_dict = {}
 
     if config_file is not None:
-        config_dict = load_yaml(config_file)
+        config_file_path = Path(config_file)
+        config_dict = load_yaml(config_file_path)
+
+        # Resolve relative paths relative to config file directory
+        config_dict = resolve_paths_relative_to_config(config_dict, config_file_path)
 
     if overrides:
         config_dict = apply_overrides(config_dict, overrides)
@@ -259,8 +345,8 @@ def load_aggregate_config(
 
 
 def load_holdout_config(
-    config_file: Optional[Union[str, Path]] = None,
-    overrides: Optional[list[str]] = None,
+    config_file: str | Path | None = None,
+    overrides: list[str] | None = None,
 ) -> HoldoutEvalConfig:
     """
     Load holdout evaluation configuration from file and CLI overrides.
@@ -275,7 +361,11 @@ def load_holdout_config(
     config_dict = {}
 
     if config_file is not None:
-        config_dict = load_yaml(config_file)
+        config_file_path = Path(config_file)
+        config_dict = load_yaml(config_file_path)
+
+        # Resolve relative paths relative to config file directory
+        config_dict = resolve_paths_relative_to_config(config_dict, config_file_path)
 
     if overrides:
         config_dict = apply_overrides(config_dict, overrides)
@@ -286,7 +376,7 @@ def load_holdout_config(
         raise ValueError(f"Invalid holdout configuration:\n{e}") from e
 
 
-def save_config(config: Union[SplitsConfig, TrainingConfig], output_path: Union[str, Path]):
+def save_config(config: SplitsConfig | TrainingConfig, output_path: str | Path):
     """Save resolved configuration to YAML file."""
     output_path = Path(output_path)
     output_path.parent.mkdir(parents=True, exist_ok=True)
@@ -310,7 +400,7 @@ def save_config(config: Union[SplitsConfig, TrainingConfig], output_path: Union[
         yaml.dump(config_dict, f, default_flow_style=False, sort_keys=False)
 
 
-def print_config_summary(config: Union[SplitsConfig, TrainingConfig], logger=None):
+def print_config_summary(config: SplitsConfig | TrainingConfig, logger=None):
     """Print human-readable configuration summary."""
     lines = []
     lines.append("=" * 80)
