@@ -10,10 +10,8 @@ References:
 - XGBoost tree_method controls CPU vs GPU acceleration
 """
 
-import math
-import numbers
 import re
-from typing import Any, Dict, List, Optional, Sequence, Tuple
+from typing import Any
 
 import numpy as np
 import sklearn
@@ -36,7 +34,7 @@ from ..config.schema import TrainingConfig
 # ----------------------------
 # sklearn version compatibility
 # ----------------------------
-def _sklearn_version_tuple(ver: str) -> Tuple[int, int, int]:
+def _sklearn_version_tuple(ver: str) -> tuple[int, int, int]:
     """Parse sklearn version string (robust to rc/dev suffixes)."""
     nums = re.findall(r"\d+", ver)
     nums = (nums + ["0", "0", "0"])[:3]
@@ -49,7 +47,7 @@ SKLEARN_VER = _sklearn_version_tuple(getattr(sklearn, "__version__", "0.0.0"))
 # ----------------------------
 # Parameter grid utilities
 # ----------------------------
-def _parse_float_list(s: str) -> List[float]:
+def _parse_float_list(s: str) -> list[float]:
     """Parse comma-separated float values."""
     if not s:
         return []
@@ -65,7 +63,7 @@ def _parse_float_list(s: str) -> List[float]:
     return out
 
 
-def _parse_int_list(s: str) -> List[int]:
+def _parse_int_list(s: str) -> list[int]:
     """Parse comma-separated integer values."""
     if not s:
         return []
@@ -81,7 +79,7 @@ def _parse_int_list(s: str) -> List[int]:
     return out
 
 
-def _require_int_list(s: str, name: str) -> List[int]:
+def _require_int_list(s: str, name: str) -> list[int]:
     """Parse and validate non-empty integer list."""
     values = _parse_int_list(s)
     if not values:
@@ -89,7 +87,7 @@ def _require_int_list(s: str, name: str) -> List[int]:
     return values
 
 
-def _parse_none_int_float_list(s: str) -> List:
+def _parse_none_int_float_list(s: str) -> list:
     """Parse list with mixed types: None, int, float, or strings like 'sqrt'."""
     if not s:
         return []
@@ -120,7 +118,7 @@ def _parse_none_int_float_list(s: str) -> List:
     return out
 
 
-def _coerce_int_or_none_list(vals: List[Any], *, name: str) -> List[Any]:
+def _coerce_int_or_none_list(vals: list[Any], *, name: str) -> list[Any]:
     """Coerce values to int or None (sklearn max_depth parameter).
 
     Accepts:
@@ -131,7 +129,7 @@ def _coerce_int_or_none_list(vals: List[Any], *, name: str) -> List[Any]:
     Raises:
         ValueError: For non-integer floats or invalid strings
     """
-    out: List[Any] = []
+    out: list[Any] = []
     for v in vals:
         if v is None:
             out.append(None)
@@ -169,8 +167,8 @@ def _coerce_int_or_none_list(vals: List[Any], *, name: str) -> List[Any]:
 
 
 def _coerce_min_samples_leaf_list(
-    vals: List[Any], *, name: str = "rf_min_samples_leaf_grid"
-) -> List[Any]:
+    vals: list[Any], *, name: str = "rf_min_samples_leaf_grid"
+) -> list[Any]:
     """Coerce min_samples_leaf grid to sklearn-compatible types.
 
     sklearn accepts:
@@ -179,7 +177,7 @@ def _coerce_min_samples_leaf_list(
 
     Also coerces whole-number floats (e.g., 5.0 -> 5) for CLI robustness.
     """
-    out: List[Any] = []
+    out: list[Any] = []
     for v in vals:
         if isinstance(v, bool):
             raise ValueError(f"{name}: invalid boolean value {v}")
@@ -227,7 +225,7 @@ def _coerce_min_samples_leaf_list(
     return out
 
 
-def parse_class_weight_options(s: str) -> List:
+def parse_class_weight_options(s: str) -> list:
     """Parse class_weight options.
 
     Examples:
@@ -260,7 +258,7 @@ def parse_class_weight_options(s: str) -> List:
 
 
 def make_logspace(
-    minv: float, maxv: float, points: int, rng: Optional[np.random.Generator] = None
+    minv: float, maxv: float, points: int, rng: np.random.Generator | None = None
 ) -> np.ndarray:
     """Generate log-spaced values for regularization parameters.
 
@@ -286,158 +284,6 @@ def make_logspace(
         samples = rng.uniform(a, b, size=points)
         return np.power(10.0, samples)
     return np.logspace(a, b, points)
-
-
-def _randomize_numeric_list(
-    values: Sequence[Any],
-    rng: Optional[np.random.Generator],
-    *,
-    as_int: bool = False,
-    min_int: Optional[int] = None,
-    max_int: Optional[int] = None,
-    unique_int: bool = False,
-    log_scale: bool = False,
-    min_float: Optional[float] = None,
-    max_float: Optional[float] = None,
-    perturb_mode: bool = False,
-) -> List[Any]:
-    """Randomize numeric grid values (used with --grid_randomize).
-
-    Non-numeric entries are preserved.
-
-    Args:
-        values: Grid values to randomize
-        rng: Random state for reproducibility
-        as_int: Coerce to integers
-        min_int/max_int: Bounds for integer values
-        unique_int: Remove duplicates in integer grids
-        log_scale: Use log-scale sampling
-        min_float/max_float: Bounds for float values
-        perturb_mode: Add noise to each grid point individually
-
-    Returns:
-        Randomized grid values
-    """
-    values_list = list(values)
-    if rng is None or len(values_list) <= 1:
-        return values_list
-
-    numeric_idx = [i for i, v in enumerate(values_list) if isinstance(v, numbers.Real)]
-    if len(numeric_idx) <= 1:
-        return values_list
-
-    numeric_vals = [float(values_list[i]) for i in numeric_idx]
-    low = min(numeric_vals)
-    high = max(numeric_vals)
-
-    if math.isclose(low, high):
-        return values_list
-
-    n = len(numeric_idx)
-    sampled_vals: List[Any]
-
-    # Perturb mode: add noise around each grid point
-    if perturb_mode and not as_int:
-        sorted_vals = sorted(numeric_vals)
-        perturbed = []
-        for i, val in enumerate(sorted_vals):
-            # Determine perturbation range based on neighbors
-            if i == 0:
-                gap = (
-                    (sorted_vals[i + 1] - val) / 2
-                    if i + 1 < len(sorted_vals)
-                    else (high - low) * 0.1
-                )
-                noise = rng.uniform(-gap * 0.5, gap)
-            elif i == len(sorted_vals) - 1:
-                gap = (val - sorted_vals[i - 1]) / 2
-                noise = rng.uniform(-gap, gap * 0.5)
-            else:
-                gap_left = (val - sorted_vals[i - 1]) / 2
-                gap_right = (sorted_vals[i + 1] - val) / 2
-                noise = rng.uniform(-gap_left, gap_right)
-
-            perturbed_val = val + noise
-            if min_float is not None:
-                perturbed_val = max(perturbed_val, float(min_float))
-            if max_float is not None:
-                perturbed_val = min(perturbed_val, float(max_float))
-            perturbed.append(perturbed_val)
-
-        sampled_vals = perturbed
-    elif as_int:
-        if perturb_mode:
-            sorted_vals = sorted(numeric_vals)
-            perturbed = []
-            for i, val in enumerate(sorted_vals):
-                int_val = int(val)
-                if i == 0:
-                    gap = (
-                        int((sorted_vals[i + 1] - val) / 2)
-                        if i + 1 < len(sorted_vals)
-                        else max(1, int((high - low) * 0.1))
-                    )
-                    offset = rng.randint(-max(1, gap // 2), gap + 1)
-                elif i == len(sorted_vals) - 1:
-                    gap = int((val - sorted_vals[i - 1]) / 2)
-                    offset = rng.randint(-gap, max(1, gap // 2) + 1)
-                else:
-                    gap_left = int((val - sorted_vals[i - 1]) / 2)
-                    gap_right = int((sorted_vals[i + 1] - val) / 2)
-                    offset = rng.randint(-gap_left, gap_right + 1)
-
-                perturbed_val = int_val + offset
-                if min_int is not None:
-                    perturbed_val = max(perturbed_val, int(min_int))
-                if max_int is not None:
-                    perturbed_val = min(perturbed_val, int(max_int))
-                perturbed.append(perturbed_val)
-
-            if unique_int:
-                perturbed = list(dict.fromkeys(perturbed))
-                while len(perturbed) < n:
-                    extra = rng.randint(low, high + 1)
-                    if min_int is not None:
-                        extra = max(extra, int(min_int))
-                    if max_int is not None:
-                        extra = min(extra, int(max_int))
-                    if extra not in perturbed:
-                        perturbed.append(extra)
-
-            sampled_vals = perturbed
-        else:
-            # Uniform sampling for integers
-            low_int = int(low) if min_int is None else int(min_int)
-            high_int = int(high) if max_int is None else int(max_int)
-            if unique_int:
-                sampled_vals = list(rng.choice(range(low_int, high_int + 1), size=n, replace=False))
-            else:
-                sampled_vals = [int(rng.randint(low_int, high_int + 1)) for _ in range(n)]
-    else:
-        # Float sampling
-        if log_scale:
-            if low <= 0 or high <= 0:
-                sampled_vals = [float(v) for v in numeric_vals]
-            else:
-                log_low = np.log10(low)
-                log_high = np.log10(high)
-                samples = rng.uniform(log_low, log_high, size=n)
-                sampled_vals = [float(np.power(10.0, s)) for s in samples]
-        else:
-            sampled_vals = [float(rng.uniform(low, high)) for _ in range(n)]
-
-        # Apply bounds
-        if min_float is not None:
-            sampled_vals = [max(float(v), float(min_float)) for v in sampled_vals]
-        if max_float is not None:
-            sampled_vals = [min(float(v), float(max_float)) for v in sampled_vals]
-
-    # Replace numeric values in original positions
-    out = list(values_list)
-    for idx, sampled_val in zip(numeric_idx, sampled_vals):
-        out[idx] = sampled_val
-
-    return out
 
 
 def compute_scale_pos_weight_from_y(y: np.ndarray) -> float:
@@ -520,11 +366,11 @@ def build_linear_svm_calibrated(
 
 def build_random_forest(
     n_estimators: int = 500,
-    max_depth: Optional[int] = None,
+    max_depth: int | None = None,
     min_samples_leaf: int = 5,
     min_samples_split: int = 2,
     max_features: str = "sqrt",
-    max_samples: Optional[float] = None,
+    max_samples: float | None = None,
     bootstrap: bool = True,
     random_state: int = 42,
     n_jobs: int = 1,
@@ -724,246 +570,5 @@ def build_models(
 
 
 # ----------------------------
-# Hyperparameter grids
+# Hyperparameter grids (moved to hyperparams.py)
 # ----------------------------
-def get_param_distributions(
-    model_name: str,
-    config: TrainingConfig,
-    feature_select: str,
-    k_grid: List[int],
-    kbest_scope: str,
-    xgb_scale_pos_weight: Optional[float] = None,
-    grid_rng: Optional[np.random.Generator] = None,
-    randomize_grids: bool = False,
-) -> Dict[str, List]:
-    """Generate hyperparameter distributions for RandomizedSearchCV.
-
-    Args:
-        model_name: Model identifier
-        config: Training configuration
-        feature_select: 'kbest', 'hybrid', or 'none'
-        k_grid: Feature selection grid
-        kbest_scope: 'protein' or 'transformed'
-        xgb_scale_pos_weight: Class balancing for XGBoost
-        grid_rng: RNG for grid randomization
-        randomize_grids: Whether to randomize grids
-
-    Returns:
-        Dictionary of parameter name -> value list
-    """
-    d = {}
-    rng = grid_rng if randomize_grids else None
-
-    # Feature selection grids
-    if feature_select in ("kbest", "hybrid"):
-        if not k_grid:
-            raise ValueError("feature_select in {kbest,hybrid} requires k_grid")
-        if kbest_scope == "protein":
-            d["prot_sel__k"] = k_grid
-        else:
-            d["sel__k"] = k_grid
-
-    # Get hyperparameter grids from config
-    # LR and SVM use C_min/C_max/C_points to generate logspace grids
-    lr_Cs = (
-        make_logspace(config.lr.C_min, config.lr.C_max, config.lr.C_points, rng=rng)
-        if hasattr(config.lr, "C_min")
-        else [0.001, 0.01, 0.1, 1.0, 10.0]
-    )
-    svm_Cs = (
-        make_logspace(config.svm.C_min, config.svm.C_max, config.svm.C_points, rng=rng)
-        if hasattr(config.svm, "C_min")
-        else [0.01, 0.1, 1.0, 10.0]
-    )
-
-    # Convert class_weight_options string to list
-    lr_class_weight = parse_class_weight_options(
-        config.lr.class_weight_options if hasattr(config.lr, "class_weight_options") else "balanced"
-    )
-    svm_class_weight = parse_class_weight_options(
-        config.svm.class_weight_options
-        if hasattr(config.svm, "class_weight_options")
-        else "balanced"
-    )
-    rf_class_weight = parse_class_weight_options(
-        config.rf.class_weight_options if hasattr(config.rf, "class_weight_options") else "balanced"
-    )
-
-    # Model-specific grids
-    if model_name == "LR_L1":
-        d.update({"clf__C": lr_Cs, "clf__class_weight": lr_class_weight})
-        return d
-
-    if model_name == "LR_EN":
-        l1_grid = (
-            config.lr.l1_ratio
-            if hasattr(config.lr, "l1_ratio")
-            else [0.05, 0.1, 0.2, 0.4, 0.6, 0.8, 0.95]
-        )
-        if rng is not None:
-            l1_grid = _randomize_numeric_list(
-                l1_grid, rng, min_float=0.0, max_float=1.0, perturb_mode=True
-            )
-        d.update(
-            {
-                "clf__C": lr_Cs,
-                "clf__l1_ratio": l1_grid,
-                "clf__class_weight": lr_class_weight,
-            }
-        )
-        return d
-
-    if model_name == "LinSVM_cal":
-        # Handle sklearn API differences
-        est_key = (
-            "estimator"
-            if "estimator" in CalibratedClassifierCV(LinearSVC()).get_params()
-            else "base_estimator"
-        )
-        d.update(
-            {
-                f"clf__{est_key}__C": svm_Cs,
-                f"clf__{est_key}__class_weight": svm_class_weight,
-            }
-        )
-        return d
-
-    if model_name == "RF":
-        n_estimators_grid = (
-            config.rf.n_estimators_grid
-            if hasattr(config.rf, "n_estimators_grid")
-            else [100, 300, 500]
-        )
-        if rng is not None:
-            n_estimators_grid = _randomize_numeric_list(
-                n_estimators_grid,
-                rng,
-                as_int=True,
-                min_int=1,
-                unique_int=True,
-                perturb_mode=True,
-            )
-
-        max_depth = (
-            config.rf.max_depth_grid if hasattr(config.rf, "max_depth_grid") else [None, 10, 20, 40]
-        )
-        if rng is not None:
-            max_depth = _randomize_numeric_list(
-                max_depth,
-                rng,
-                as_int=True,
-                min_int=1,
-                unique_int=True,
-                perturb_mode=True,
-            )
-
-        min_leaf = (
-            config.rf.min_samples_leaf_grid
-            if hasattr(config.rf, "min_samples_leaf_grid")
-            else [1, 2, 4]
-        )
-        min_split = (
-            config.rf.min_samples_split_grid
-            if hasattr(config.rf, "min_samples_split_grid")
-            else [2, 5, 10]
-        )
-        max_feat = (
-            config.rf.max_features_grid
-            if hasattr(config.rf, "max_features_grid")
-            else ["sqrt", 0.5]
-        )
-
-        if rng is not None:
-            max_feat = _randomize_numeric_list(
-                max_feat, rng, min_float=0.1, max_float=1.0, perturb_mode=True
-            )
-            min_split = _randomize_numeric_list(
-                min_split,
-                rng,
-                as_int=True,
-                min_int=2,
-                unique_int=True,
-                perturb_mode=True,
-            )
-
-        d.update(
-            {
-                "clf__n_estimators": n_estimators_grid,
-                "clf__max_depth": max_depth,
-                "clf__min_samples_leaf": min_leaf,
-                "clf__min_samples_split": min_split,
-                "clf__max_features": max_feat,
-                "clf__class_weight": rf_class_weight,
-            }
-        )
-
-        return d
-
-    if model_name == "XGBoost":
-        n_estimators_grid = (
-            config.xgboost.n_estimators_grid
-            if hasattr(config.xgboost, "n_estimators_grid")
-            else [100, 300, 500]
-        )
-        max_depth_grid = (
-            config.xgboost.max_depth_grid
-            if hasattr(config.xgboost, "max_depth_grid")
-            else [3, 5, 7]
-        )
-        learning_rate_grid = (
-            config.xgboost.learning_rate_grid
-            if hasattr(config.xgboost, "learning_rate_grid")
-            else [0.01, 0.05, 0.1]
-        )
-        subsample_grid = (
-            config.xgboost.subsample_grid
-            if hasattr(config.xgboost, "subsample_grid")
-            else [0.7, 0.8, 1.0]
-        )
-        colsample_grid = (
-            config.xgboost.colsample_bytree_grid
-            if hasattr(config.xgboost, "colsample_bytree_grid")
-            else [0.7, 0.8, 1.0]
-        )
-        spw_grid = [float(xgb_scale_pos_weight)] if xgb_scale_pos_weight is not None else [1.0]
-
-        if rng is not None:
-            n_estimators_grid = _randomize_numeric_list(
-                n_estimators_grid, rng, as_int=True, min_int=1, perturb_mode=True
-            )
-            max_depth_grid = _randomize_numeric_list(
-                max_depth_grid,
-                rng,
-                as_int=True,
-                min_int=1,
-                unique_int=True,
-                perturb_mode=True,
-            )
-            learning_rate_grid = _randomize_numeric_list(
-                learning_rate_grid,
-                rng,
-                min_float=1e-4,
-                log_scale=True,
-                perturb_mode=True,
-            )
-            subsample_grid = _randomize_numeric_list(
-                subsample_grid, rng, min_float=0.1, max_float=1.0, perturb_mode=True
-            )
-            colsample_grid = _randomize_numeric_list(
-                colsample_grid, rng, min_float=0.1, max_float=1.0, perturb_mode=True
-            )
-            spw_grid = _randomize_numeric_list(spw_grid, rng, min_float=1e-3, perturb_mode=True)
-
-        d.update(
-            {
-                "clf__n_estimators": n_estimators_grid,
-                "clf__max_depth": max_depth_grid,
-                "clf__learning_rate": learning_rate_grid,
-                "clf__subsample": subsample_grid,
-                "clf__colsample_bytree": colsample_grid,
-                "clf__scale_pos_weight": spw_grid,
-            }
-        )
-        return d
-
-    return {}
