@@ -5,8 +5,8 @@ Provides functions to generate combined plots across CV repeats
 with confidence bands showing variability across splits.
 """
 
+from collections.abc import Sequence
 from pathlib import Path
-from typing import Sequence
 
 import numpy as np
 
