@@ -118,16 +118,16 @@ class TestSaveSplitsDevelopmentMode:
 
         assert result.exit_code == 0, f"CLI failed: {result.output}"
 
-        # Check output files exist
-        assert (outdir / "train_idx_seed0.csv").exists()
-        assert (outdir / "val_idx_seed0.csv").exists()
-        assert (outdir / "test_idx_seed0.csv").exists()
-        assert (outdir / "split_meta_seed0.json").exists()
+        # Check output files exist (new format with scenario in filename)
+        assert (outdir / "train_idx_IncidentOnly_seed0.csv").exists()
+        assert (outdir / "val_idx_IncidentOnly_seed0.csv").exists()
+        assert (outdir / "test_idx_IncidentOnly_seed0.csv").exists()
+        assert (outdir / "split_meta_IncidentOnly_seed0.json").exists()
 
         # Validate splits
-        train_idx = pd.read_csv(outdir / "train_idx_seed0.csv")["idx"].values
-        val_idx = pd.read_csv(outdir / "val_idx_seed0.csv")["idx"].values
-        test_idx = pd.read_csv(outdir / "test_idx_seed0.csv")["idx"].values
+        train_idx = pd.read_csv(outdir / "train_idx_IncidentOnly_seed0.csv")["idx"].values
+        val_idx = pd.read_csv(outdir / "val_idx_IncidentOnly_seed0.csv")["idx"].values
+        test_idx = pd.read_csv(outdir / "test_idx_IncidentOnly_seed0.csv")["idx"].values
 
         # No overlap
         assert len(set(train_idx) & set(val_idx)) == 0
@@ -140,7 +140,7 @@ class TestSaveSplitsDevelopmentMode:
         assert all(test_idx >= 0)
 
         # Validate metadata
-        with open(outdir / "split_meta_seed0.json") as f:
+        with open(outdir / "split_meta_IncidentOnly_seed0.json") as f:
             meta = json.load(f)
 
         assert meta["scenario"] == "IncidentOnly"
@@ -185,15 +185,15 @@ class TestSaveSplitsDevelopmentMode:
 
         assert result.exit_code == 0, f"CLI failed: {result.output}"
 
-        # Files should exist
-        assert (outdir / "train_idx_seed0.csv").exists()
-        assert (outdir / "val_idx_seed0.csv").exists()
-        assert (outdir / "test_idx_seed0.csv").exists()
+        # Files should exist (new format with scenario in filename)
+        assert (outdir / "train_idx_IncidentPlusPrevalent_seed0.csv").exists()
+        assert (outdir / "val_idx_IncidentPlusPrevalent_seed0.csv").exists()
+        assert (outdir / "test_idx_IncidentPlusPrevalent_seed0.csv").exists()
 
         # Load original data to verify prevalent handling
         df = pd.read_csv(toy_proteomics_csv)
-        val_idx = pd.read_csv(outdir / "val_idx_seed0.csv")["idx"].values
-        test_idx = pd.read_csv(outdir / "test_idx_seed0.csv")["idx"].values
+        val_idx = pd.read_csv(outdir / "val_idx_IncidentPlusPrevalent_seed0.csv")["idx"].values
+        test_idx = pd.read_csv(outdir / "test_idx_IncidentPlusPrevalent_seed0.csv")["idx"].values
 
         # VAL and TEST should only have Incident + Controls (no prevalent)
         val_labels = df.loc[val_idx, TARGET_COL].values
@@ -237,7 +237,7 @@ class TestSaveSplitsDevelopmentMode:
 
         # Check that TRAIN has reduced controls
         df = pd.read_csv(toy_proteomics_csv)
-        train_idx = pd.read_csv(outdir / "train_idx_seed0.csv")["idx"].values
+        train_idx = pd.read_csv(outdir / "train_idx_IncidentOnly_seed0.csv")["idx"].values
         train_labels = df.loc[train_idx, TARGET_COL].values
 
         n_cases = (train_labels == INCIDENT_LABEL).sum()
@@ -279,16 +279,16 @@ class TestSaveSplitsDevelopmentMode:
 
         assert result.exit_code == 0, f"CLI failed: {result.output}"
 
-        # Check that all 3 splits were created
+        # Check that all 3 splits were created (new format with scenario in filename)
         for seed in [100, 101, 102]:
-            assert (outdir / f"train_idx_seed{seed}.csv").exists()
-            assert (outdir / f"val_idx_seed{seed}.csv").exists()
-            assert (outdir / f"test_idx_seed{seed}.csv").exists()
-            assert (outdir / f"split_meta_seed{seed}.json").exists()
+            assert (outdir / f"train_idx_IncidentOnly_seed{seed}.csv").exists()
+            assert (outdir / f"val_idx_IncidentOnly_seed{seed}.csv").exists()
+            assert (outdir / f"test_idx_IncidentOnly_seed{seed}.csv").exists()
+            assert (outdir / f"split_meta_IncidentOnly_seed{seed}.json").exists()
 
         # Verify splits are different
-        train_100 = pd.read_csv(outdir / "train_idx_seed100.csv")["idx"].values
-        train_101 = pd.read_csv(outdir / "train_idx_seed101.csv")["idx"].values
+        train_100 = pd.read_csv(outdir / "train_idx_IncidentOnly_seed100.csv")["idx"].values
+        train_101 = pd.read_csv(outdir / "train_idx_IncidentOnly_seed101.csv")["idx"].values
 
         # Different seeds should produce different splits
         assert not np.array_equal(sorted(train_100), sorted(train_101))
@@ -347,16 +347,16 @@ class TestSaveSplitsHoldoutMode:
         assert (outdir / "HOLDOUT_idx.csv").exists()
         assert (outdir / "HOLDOUT_meta.json").exists()
 
-        # Check development split files
-        assert (outdir / "train_idx_seed0.csv").exists()
-        assert (outdir / "val_idx_seed0.csv").exists()
-        assert (outdir / "test_idx_seed0.csv").exists()
+        # Check development split files (new format with scenario in filename)
+        assert (outdir / "train_idx_IncidentOnly_seed0.csv").exists()
+        assert (outdir / "val_idx_IncidentOnly_seed0.csv").exists()
+        assert (outdir / "test_idx_IncidentOnly_seed0.csv").exists()
 
         # Load indices
         holdout_idx = pd.read_csv(outdir / "HOLDOUT_idx.csv")["idx"].values
-        train_idx = pd.read_csv(outdir / "train_idx_seed0.csv")["idx"].values
-        val_idx = pd.read_csv(outdir / "val_idx_seed0.csv")["idx"].values
-        test_idx = pd.read_csv(outdir / "test_idx_seed0.csv")["idx"].values
+        train_idx = pd.read_csv(outdir / "train_idx_IncidentOnly_seed0.csv")["idx"].values
+        val_idx = pd.read_csv(outdir / "val_idx_IncidentOnly_seed0.csv")["idx"].values
+        test_idx = pd.read_csv(outdir / "test_idx_IncidentOnly_seed0.csv")["idx"].values
 
         # Holdout should be disjoint from development sets
         dev_idx = np.concatenate([train_idx, val_idx, test_idx])
@@ -404,7 +404,8 @@ class TestSaveSplitsErrorHandling:
         )
         assert result1.exit_code == 0
 
-        # Second run without --overwrite should fail
+        # Second run without --overwrite should succeed (idempotent behavior)
+        # Refactored code allows re-running with same config (logs warning, no error)
         result2 = runner.invoke(
             cli,
             [
@@ -425,10 +426,8 @@ class TestSaveSplitsErrorHandling:
                 "0.25",
             ],
         )
-        assert result2.exit_code != 0
-        # Check for error message in exception string
-        exc_str = str(result2.exception)
-        assert "already exist" in exc_str or "FileExistsError" in exc_str
+        # Should succeed (idempotent - same config is allowed to avoid breaking workflows)
+        assert result2.exit_code == 0
 
     def test_invalid_scenario_raises(self, toy_proteomics_csv, tmp_path):
         """Test that invalid scenario name raises error."""
@@ -514,16 +513,16 @@ class TestSaveSplitsReproducibility:
         )
         assert result2.exit_code == 0
 
-        # Compare indices
-        train1 = pd.read_csv(outdir1 / "train_idx_seed42.csv")["idx"].values
-        train2 = pd.read_csv(outdir2 / "train_idx_seed42.csv")["idx"].values
+        # Compare indices (new format with scenario in filename)
+        train1 = pd.read_csv(outdir1 / "train_idx_IncidentOnly_seed42.csv")["idx"].values
+        train2 = pd.read_csv(outdir2 / "train_idx_IncidentOnly_seed42.csv")["idx"].values
 
         np.testing.assert_array_equal(train1, train2)
 
         # Compare metadata (excluding timestamps if any)
-        with open(outdir1 / "split_meta_seed42.json") as f:
+        with open(outdir1 / "split_meta_IncidentOnly_seed42.json") as f:
             meta1 = json.load(f)
-        with open(outdir2 / "split_meta_seed42.json") as f:
+        with open(outdir2 / "split_meta_IncidentOnly_seed42.json") as f:
             meta2 = json.load(f)
 
         # Check key fields match
